feat(config): add region setting to sst.config.ts (#117)
* feat(config): add region setting to sst.config.ts

Add 'us-east-1' as the default region for the SST configuration to ensure consistent deployment location

* chore: update Node.js version in .nvmrc to v20

This change updates the Node.js version specified in .nvmrc from v16 to v20 to ensure compatibility with the latest features and improvements.

* refactor: reduce query limits for tags, posts, and categories

Optimize query performance by reducing the number of items fetched in a single request. This change decreases the load on the server and improves response times.

* ci: add autodeploy configuration for production stage

Enable automatic deployment to the production stage when changes are pushed to the main branch. This ensures that the latest updates are deployed without manual intervention.